requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3
Django==2.0.7
idna==2.7
mongoengine==0.15.3
pycparser==2.18
PyJWT==1.6.4
pymongo==3.7.1
pytz==2018.5
six==1.11.0
